--- src/main/java/frc/robot/subsystems/ResetEncoderPosition.java
@@ -1,18 +1,0 @@
-package frc.robot.subsystems;
-
-import edu.wpi.first.wpilibj.command.InstantCommand;
-
-public class ResetEncoderPosition extends InstantCommand {
-
-    private Arm arm;
-
-    public ResetEncoderPosition(Arm arm) {
-        requires(arm);
-        this.arm = arm;
-    }
-
-    @Override
-    protected void initialize() {
-        arm.resetElevatorPosition();
-    }
-}
